+++ tests/test_c3_parser.py
@@ -1,7 +1,7 @@
 import datetime
 import unittest
 
-from app.parser.target import C3NearTargetsParser
+from app.parser.c3 import C3NearTargetsParser
 from tests.base import read_input_img
 
 
@@ -9,14 +9,23 @@ class TestNearestTargets(unittest.TestCase):
     def setUp(self):
         self.parser = C3NearTargetsParser(debug=False)
 
-    def test_ttt(self):
+    def test_parsing1(self):
         rgb, grey = read_input_img("../res/input/c3/Shot00004.bmp")
 
-        print("Start %s" % datetime.datetime.now())
+        print("Start %s\n" % datetime.datetime.now())
         result = self.parser.parse_image(rgb, grey)
         for r in result:
             print("%s distance -> %s" % (r.name, r.distance))
-        print("Finish %s" % datetime.datetime.now())
+        print("\nFinish %s\n" % datetime.datetime.now())
+
+    def test_parsing2(self):
+        rgb, grey = read_input_img("../res/input/c3/Shot00008.bmp")
+
+        print("Start %s\n" % datetime.datetime.now())
+        result = self.parser.parse_image(rgb, grey)
+        for r in result:
+            print("%s distance -> %s" % (r.name, r.distance))
+        print("\nFinish %s" % datetime.datetime.now())
 
 
 if __name__ == "__main__":
